+++ test/test_helper.rb
@@ -32,6 +32,7 @@ module Studio
   mattr_accessor :sso_logo,            default: nil
   mattr_accessor :wallet_address_method, default: nil
   mattr_accessor :theme_logos,         default: []
+  mattr_accessor :sticky_table_headers, default: false
 
   mattr_accessor :theme_primary,  default: "#8E82FE"
   mattr_accessor :theme_dark,     default: "#1A1535"